+++ usipipo_commons/domain/entities/__init__.py
@@ -1,5 +1,6 @@
 from .consumption_billing import ConsumptionBilling
 from .consumption_invoice import ConsumptionInvoice
+from .crypto_order import CryptoOrder
 from .payment import Payment
 from .user import User
 from .vpn_key import VpnKey
@@ -10,4 +11,5 @@ __all__ = [
     "Payment",
     "ConsumptionBilling",
     "ConsumptionInvoice",
+    "CryptoOrder",
 ]